♻️ #33 elasticSearch 리팩토링
