+++ backend/schema.gql
@@ -59,7 +59,6 @@ type Query {
   crawling_test: String!
   aniType(type: String!): [Animation!]!
   searchTitle(title: String!): [Animation!]!
-  elasticTest: String!
   readReivew: [Review!]!
   detailReview(animationId: Int!, page: Int!, pageSize: Int!): UserReviewResponse!
   allAnimations: [Animation!]!
@@ -69,6 +68,8 @@ type Query {
   similarAnimation(id: Int!): [Animation!]!
   readWishList: [Animation!]!
   wishYN(animationId: Float!): Boolean!
+  settingAnalyzer: String!
+  indexAnimation: String!
 }
 
 type Mutation {
