+++ backend/src/app.module.ts
@@ -12,6 +12,7 @@ import { ReviewModule } from './review/review.module';
 import { AnimationModule } from './animation/animation.module';
 import { WishModule } from './wish/wish.module';
 import { ElasticsearchModule } from '@nestjs/elasticsearch';
+import { MyElasticSearchModule } from './elasticSearch/elasticSearch.module';
 @Module({
   imports: [
     ConfigModule.forRoot(),
@@ -28,7 +29,8 @@ import { ElasticsearchModule } from '@nestjs/elasticsearch';
     SearchModule,
     ReviewModule,
     AnimationModule,
-    WishModule
+    WishModule,
+    MyElasticSearchModule
   ],
   providers: [AppResolver],
 })
